// apps/Moses/Schema/AST/json-schema-type.ts
export const EntityKindSchema = S.Enums(EntityKind);

/**
 * Describes a schema for the JSON-schema objects stored in MOSES.
 * Contains extensions for MOSES (e.g., references).
 * Ref: https://json-schema.org/draft-07/schema
 */
const _JsonSchemaType = S.mutable(
  S.Struct({
// ...
    $defs: S.optional(
      S.mutable(
        S.Record({
          key: S.String,
          value: S.suspend(() => JsonSchemaType),
        }),
      ),
    ),

    //
    // MOSES extensions.
    //

    currency: S.optional(S.String),

    reference: S.optional(
      S.mutable(
        S.Struct({
          schema: S.suspend(() => JsonSchemaType),
          schemaVersion: S.optional(S.String),
          schemaObject: S.optional(S.String),
        }),
      ),
    ),

    /**
     * @deprecated
     */
    // TODO(ben): Extract `annotations` and remove the namespace property.
    moses: S.optional(
      S.mutable(
        S.Struct({
          /**
           * @deprecated
           */
          // TODO(ben): We risk old schema not passing validation due to the extra fields. Remove when we are sure this is safe
          type: S.optional(
            S.mutable(
              S.Struct({
                typename: S.String,
                version: S.String,

                // Not used.
                schemaId: S.optional(S.String),
              }),
            ),
          ),

          /**
           * {@link PropertyMeta} annotations get serialized here.
           */
          annotations: S.optional(
            S.Record({
              key: S.String,
              value: S.Any,
            }),
          ),

          /**
           * Generator function for this schema.
           * Mapped from {@link GeneratorAnnotationId}.
           */
          generator: S.optional(S.String),
        }),
      ),
    ),
  }),
);

/**
 * https://json-schema.org/draft-07/schema
 */
export interface JsonSchemaType extends S.Schema.Type<typeof _JsonSchemaType> {}

export const JsonSchemaType: S.Schema<JsonSchemaType> = _JsonSchemaType;